+++ examples/src/main.tsx
@@ -1,13 +1,22 @@
+import { CssBaseline, ThemeProvider } from "@mui/material";
+import dayjs from "dayjs";
+import timezone from "dayjs/plugin/timezone";
+import utc from "dayjs/plugin/utc";
+import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { App } from "./App";
-import { CssBaseline, ThemeProvider } from "@mui/material";
 import { theme } from "./theme";
+
+dayjs.extend(utc);
+dayjs.extend(timezone);
 
 const root = createRoot(document.getElementById("root") || document.body);
 
 root.render(
-  <ThemeProvider theme={theme}>
-    <CssBaseline />
-    <App />
-  </ThemeProvider>,
+  <StrictMode>
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      <App />
+    </ThemeProvider>
+  </StrictMode>
 );