+++ examples/src/samples.ts
@@ -1,0 +1,27 @@
+const githubSamplesLocation =
+  "https://github.com/ukorvl/lightweight-charts-react-components/tree/main/examples/src/samples";
+
+const samplesLinks = {
+  BasicSeries: {
+    githbub: `${githubSamplesLocation}/BasicSeries`,
+    codesandbox: "",
+  },
+  CustomSeries: {
+    githbub: `${githubSamplesLocation}/CustomSeries`,
+    codesandbox: "",
+  },
+  RangeSwitcher: {
+    githbub: `${githubSamplesLocation}/RangeSwitcher`,
+    codesandbox: "",
+  },
+  Markers: {
+    github: `${githubSamplesLocation}/Markers`,
+    codesandbox: "",
+  },
+  Watermark: {
+    github: `${githubSamplesLocation}/Watermark`,
+    codesandbox: "",
+  },
+} as const;
+
+export { samplesLinks };