+++ examples/src/samples/CustomSeries/CustomSeries.tsx
@@ -1,16 +1,18 @@
-import { ChartWidgetCard } from "@/ui/ChartWidgetCard";
 import {
   Chart,
   CustomSeries as CustomSeriesComponent,
 } from "lightweight-charts-react-components";
-import { GroupedBarsSeries } from "./plugin";
 import { chartCommonOptions } from "@/common/chartCommonOptions";
+import { samplesLinks } from "@/samples";
+import { ChartWidgetCard } from "@/ui/ChartWidgetCard";
+import { GroupedBarsSeries } from "./plugin";
 
 const CustomSeries = () => {
   return (
     <ChartWidgetCard
       title="Custom series"
       subTitle="Custom series plugin usage (grouped bars)"
+      githubLink={samplesLinks.CustomSeries.githbub}
     >
       <Chart height={400} {...chartCommonOptions} autoSize>
         <CustomSeriesComponent