+++ examples/src/samples/RangeSwitcher/RangeSwitcher.tsx
@@ -1,18 +1,19 @@
-import { chartCommonOptions } from "@/common/chartCommonOptions";
-import { ChartWidgetCard } from "@/ui/ChartWidgetCard";
 import { Button, ButtonGroup } from "@mui/material";
 import { AreaSeries, Chart } from "lightweight-charts-react-components";
+import { colors } from "@/colors";
+import { chartCommonOptions } from "@/common/chartCommonOptions";
+import { typedObjectKeys } from "@/common/utils";
+import { samplesLinks } from "@/samples";
+import { ChartWidgetCard } from "@/ui/ChartWidgetCard";
 import {
   dataRangeMap,
   useDataRangeStore,
   useSeriesDataStore,
 } from "./rangeSwitcherStore";
-import { typedObjectKeys } from "@/common/utils";
-import { AreaSeriesOptions, DeepPartial } from "lightweight-charts";
-import { colors } from "@/colors";
+import type { AreaSeriesOptions, DeepPartial } from "lightweight-charts";
 
 const seriesCustomOptions = {
-  bottomColor: `${colors.blue}05`,
+  bottomColor: `${colors.blue100}05`,
   lineColor: colors.cyan,
   topColor: colors.cyan,
 } satisfies DeepPartial<AreaSeriesOptions>;
@@ -25,13 +26,15 @@ const RangeSwitcher = () => {
     <ChartWidgetCard
       title="Range switcher"
       subTitle="Allows user to switch between different time ranges"
+      githubLink={samplesLinks.RangeSwitcher.githbub}
     >
       <ButtonGroup
         variant="outlined"
         aria-label="Basic button group"
         color="info"
+        sx={{ marginBottom: 2 }}
       >
-        {typedObjectKeys(dataRangeMap).map((key) => (
+        {typedObjectKeys(dataRangeMap).map(key => (
           <Button
             key={key}
             onClick={() => setRange(key)}
@@ -48,8 +51,9 @@ const RangeSwitcher = () => {
         }}
         {...chartCommonOptions}
         autoSize
+        onInit={chart => chart.timeScale().fitContent()}
       >
-        <AreaSeries reactive options={seriesCustomOptions} data={data} />
+        <AreaSeries options={seriesCustomOptions} data={data} />
       </Chart>
     </ChartWidgetCard>
   );