+++ examples/src/theme.ts
@@ -3,6 +3,22 @@ import { colors } from "./colors";
 
 const theme = createTheme({
   cssVariables: true,
+  components: {
+    MuiCardHeader: {
+      styleOverrides: {
+        action: {
+          margin: 0,
+        },
+      },
+    },
+    MuiSwitch: {
+      styleOverrides: {
+        track: {
+          backgroundColor: colors.white,
+        },
+      },
+    },
+  },
   palette: {
     primary: {
       main: colors.pink,
@@ -15,12 +31,14 @@ const theme = createTheme({
     },
     background: {
       default: colors.black,
-      paper: colors.darkBlue,
+      paper: colors.blue200,
     },
     text: {
       primary: colors.white,
-      secondary: colors.lightBlue,
+      secondary: colors.blue,
+      disabled: colors.gray100,
     },
+    divider: colors.gray100,
   },
   typography: {
     fontFamily: "Roboto",