+++ examples/src/ui/LayoutGrid.tsx
@@ -1,15 +1,31 @@
 import { Grid2 as Grid } from "@mui/material";
-import { Children, FC, ReactNode } from "react";
+import { Children } from "react";
+import type { FC, ReactNode } from "react";
 
 type LayoutGridProps = {
   children: ReactNode;
 };
 
 const LayoutGrid: FC<LayoutGridProps> = ({ children }) => {
+  const childrenLength = Children.count(children);
+  const numberOfChildrenIsOdd = childrenLength % 2 === 1;
+
   return (
-    <Grid container spacing={4}>
-      {Children.map(children, (child) => (
-        <Grid size={{ xs: 12, md: 12 }}>{child}</Grid>
+    <Grid
+      container
+      spacing={{ xs: 2, sm: 4 }}
+      justifyContent={numberOfChildrenIsOdd ? "center" : "stretch"}
+    >
+      {Children.map(children, (child, i) => (
+        <Grid
+          size={{
+            xs: 12,
+            lg: numberOfChildrenIsOdd ? (i === childrenLength - 1 ? 8 : 6) : 6,
+          }}
+          key={i}
+        >
+          {child}
+        </Grid>
       ))}
     </Grid>
   );