+++ lib/src/markers/Markers.tsx
@@ -1,12 +1,13 @@
-import { ForwardedRef, forwardRef, useImperativeHandle } from "react";
-import { MarkersApiRef, MarkersProps } from "./types";
-import { useInitMarkers } from "./useInitMarkers";
+import { forwardRef, useImperativeHandle } from "react";
+import { useMarkers } from "./useMarkers";
+import type { MarkersApiRef, MarkersProps } from "./types";
+import type { ForwardedRef } from "react";
 
 const MarkersRenderFunction = (
   { ...rest }: MarkersProps,
-  ref: ForwardedRef<MarkersApiRef>,
+  ref: ForwardedRef<MarkersApiRef>
 ) => {
-  const markersApiRef = useInitMarkers(rest);
+  const markersApiRef = useMarkers(rest);
   useImperativeHandle(ref, () => markersApiRef.current, [markersApiRef]);
 
   return null;