+++ lib/src/markers/types.ts
@@ -1,10 +1,10 @@
-import { ISeriesMarkersPluginApi, SeriesMarker, Time } from "lightweight-charts";
+import type { ISeriesMarkersPluginApi, SeriesMarker, Time } from "lightweight-charts";
 
 export type MarkersApiRef = {
   _markers: ISeriesMarkersPluginApi<Time> | null;
   api: () => ISeriesMarkersPluginApi<Time> | null;
+  init: () => ISeriesMarkersPluginApi<Time> | null;
   clear: () => void;
-  destroyed: boolean;
 };
 
 export type MarkersProps = {