+++ lib/src/priceLine/PriceLine.tsx
@@ -1,9 +1,13 @@
-import { ForwardedRef, forwardRef, useImperativeHandle } from "react";
-import { PriceLineApiRef, PriceLineProps } from "./types";
-import { useInitPriceLine } from "./useInitPriceLine";
+import { forwardRef, useImperativeHandle } from "react";
+import { usePriceLine } from "./usePriceLine";
+import type { PriceLineApiRef, PriceLineProps } from "./types";
+import type { ForwardedRef } from "react";
 
-const PriceLineRenderFunction = (props: PriceLineProps, ref: ForwardedRef<PriceLineApiRef>) => {
-  const priceLineApiRef = useInitPriceLine(props);
+const PriceLineRenderFunction = (
+  props: PriceLineProps,
+  ref: ForwardedRef<PriceLineApiRef>
+) => {
+  const priceLineApiRef = usePriceLine(props);
   useImperativeHandle(ref, () => priceLineApiRef.current, [priceLineApiRef]);
 
   return null;