+++ lib/src/scales/TimeScale.tsx
@@ -1,9 +1,13 @@
-import { ForwardedRef, forwardRef, useImperativeHandle } from "react";
-import { TimeScaleApiRef, TimeScaleProps } from "./types";
-import { useInitTimeScale } from "./useInitTimeScale";
+import { forwardRef, useImperativeHandle } from "react";
+import { useTimeScale } from "./useTimeScale";
+import type { TimeScaleApiRef, TimeScaleProps } from "./types";
+import type { ForwardedRef } from "react";
 
-const TimeScaleRenderFunction = (props: TimeScaleProps, ref: ForwardedRef<TimeScaleApiRef>) => {
-  const timeScaleApiRef = useInitTimeScale(props);
+const TimeScaleRenderFunction = (
+  props: TimeScaleProps,
+  ref: ForwardedRef<TimeScaleApiRef>
+) => {
+  const timeScaleApiRef = useTimeScale(props);
   useImperativeHandle(ref, () => timeScaleApiRef.current, [timeScaleApiRef]);
 
   return null;