+++ lib/src/series/useSeries.ts
@@ -1,10 +1,3 @@
-import { useLayoutEffect, useRef } from "react";
-import {
-  CustomSeriesUniqueProps,
-  SeriesApiRef,
-  SeriesTemplateProps,
-  SeriesType,
-} from "./types";
 import {
   LineSeries,
   CandlestickSeries,
@@ -12,29 +5,40 @@ import {
   AreaSeries,
   BaselineSeries,
   BarSeries,
-  SeriesDefinition,
-  ISeriesApi,
 } from "lightweight-charts";
-import { useSafeContext } from "@/shared/useSafeContext";
+import { useLayoutEffect, useRef, useState } from "react";
 import { ChartContext } from "@/chart/ChartContext";
 import { BaseInternalError } from "@/shared/InternalError";
+import { useSafeContext } from "@/shared/useSafeContext";
+import type {
+  CustomSeriesUniqueProps,
+  SeriesApiRef,
+  SeriesTemplateProps,
+  SeriesType,
+} from "./types";
+import type { SeriesDefinition, ISeriesApi } from "lightweight-charts";
 
 type SeriesTypeWithoutCustom = Exclude<SeriesType, "Custom">;
 
-export const useInitSeries = <T extends SeriesType>({
+export const useSeries = <T extends SeriesType>({
   type,
   data,
   options = {},
-  reactive,
+  reactive = true,
   ...rest
 }: Omit<SeriesTemplateProps<T>, "children">) => {
-  const chart = useSafeContext(ChartContext);
+  const { initialized: chartInitialized, chartApiRef: chart } =
+    useSafeContext(ChartContext);
+  const [initialized, setInitialized] = useState(false);
 
   const seriesApiRef = useRef<SeriesApiRef<T>>({
     _series: null,
     api() {
-      if (!this._series && !this.destroyed) {
-        const chartApi = chart.api();
+      return this._series;
+    },
+    init() {
+      if (!this._series) {
+        const chartApi = chart?.api();
 
         if (!chartApi) {
           return null;
@@ -49,14 +53,16 @@ export const useInitSeries = <T extends SeriesType>({
           // TODO: Fix this type cast and infer the correct type
           (this._series as unknown as ISeriesApi<"Custom">) = chartApi.addCustomSeries(
             plugin,
-            options,
+            options
           );
         } else {
           this._series = chartApi.addSeries(
             seriesMap[type as SeriesTypeWithoutCustom] as SeriesDefinition<T>,
-            options,
+            options
           );
         }
+
+        setInitialized(true);
 
         this._series?.setData(data);
       }
@@ -65,17 +71,20 @@ export const useInitSeries = <T extends SeriesType>({
     },
     clear() {
       if (this._series !== null) {
-        chart.api()?.removeSeries(this._series);
+        chart?.api()?.removeSeries(this._series);
         this._series = null;
-        this.destroyed = true;
+        setInitialized(false);
       }
     },
-    destroyed: false,
   });
 
   useLayoutEffect(() => {
-    seriesApiRef.current.api();
+    if (!chartInitialized) return;
 
+    seriesApiRef.current.init();
+  }, [chartInitialized]);
+
+  useLayoutEffect(() => {
     return () => {
       seriesApiRef.current.clear();
     };
@@ -97,7 +106,7 @@ export const useInitSeries = <T extends SeriesType>({
     }
   }, [options]);
 
-  return seriesApiRef;
+  return { initialized, seriesApiRef };
 };
 
 const seriesMap: Record<