+++ lib/src/watermark/index.ts
@@ -1,0 +1,7 @@
+export { WatermarkText, WatermarkImage } from "./Watermark";
+export type {
+  WatermarkApiRef,
+  WatermarkType,
+  TextWatermarkProps,
+  ImageWatermarkProps,
+} from "./types";