+++ lib/CHANGELOG.md
@@ -1,7 +1,24 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),  
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [0.2.0] - 2025-04-05
+### Fix 
+- fix rendering in strict mode
+
+## [0.1.2] - 2025-04-05
+### Feat
+- add WatermarkText and WatermarkImage components
+
+## [0.1.1] - 2025-04-03
+### Fix
+- fix empty markers on init issue
+### Feat
+- make series and markers reactive by default
 
 ## [0.1.0] - 2025-03-31
 ### Feat